+++ json_serializable/test/custom_configuration_test.dart
@@ -32,7 +32,7 @@ Future<void> main() async {
   );
 
   group('without wrappers', () {
-    _registerTests(ClassConfig.defaults);
+    _registerTests(ClassConfig.defaults.toJsonSerializable());
   });
 
   group('configuration', () {
@@ -61,8 +61,10 @@ Future<void> main() async {
                 nullConfig ? null : const JsonSerializable(), className);
 
             expect(_ConfigLogger.configurations, hasLength(2));
-            expect(_ConfigLogger.configurations.first,
-                same(_ConfigLogger.configurations.last));
+            expect(
+              _ConfigLogger.configurations.first.toJson(),
+              _ConfigLogger.configurations.last.toJson(),
+            );
             expect(_ConfigLogger.configurations.first.toJson(),
                 generatorConfigDefaultJson);
           });
@@ -98,8 +100,10 @@ Future<void> main() async {
             'ConfigurationExplicitDefaults');
 
         expect(_ConfigLogger.configurations, hasLength(2));
-        expect(_ConfigLogger.configurations.first,
-            same(_ConfigLogger.configurations.last));
+        expect(
+          _ConfigLogger.configurations.first.toJson(),
+          _ConfigLogger.configurations.last.toJson(),
+        );
 
         // The effective configuration should be non-Default configuration, but
         // with all fields set from JsonSerializable as the defaults
@@ -225,14 +229,14 @@ class _ConfigLogger implements TypeHelper<TypeHelperContextWithConfig> {
     TypeHelperContextWithConfig context,
     bool defaultProvided,
   ) {
-    configurations.add(context.config);
+    configurations.add(context.config.toJsonSerializable());
     return null;
   }
 
   @override
   Object? serialize(DartType targetType, String expression,
       TypeHelperContextWithConfig context) {
-    configurations.add(context.config);
+    configurations.add(context.config.toJsonSerializable());
     return null;
   }
 }