+++ json_serializable/test/kitchen_sink/kitchen_sink.dart
@@ -71,6 +71,7 @@ class _Factory implements k.KitchenSinkFactory<String, dynamic> {
         TrivialNumber(0),
         {},
         DateTime.fromMillisecondsSinceEpoch(0),
+        TrivialString(''),
         TrivialNumber(0),
         {},
       );
@@ -194,11 +195,14 @@ class KitchenSink implements k.KitchenSink {
   }
 }
 
-@JsonSerializable()
+@JsonSerializable(converters: [
+  // referencing a top-level field should work
+  durationConverter,
+  // referencing via a const constructor should work
+  BigIntStringConverter(),
+])
 // referencing a top-level field should work
-@durationConverter
-// referencing via a const constructor should work
-@BigIntStringConverter()
+@trivialStringConverter
 @TrivialNumberConverter.instance
 @EpochDateTimeConverter()
 class JsonConverterTestClass implements k.JsonConverterTestClass {
@@ -212,6 +216,7 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
     this.numberSilly,
     this.numberSillySet,
     this.dateTime,
+    this.trivialString,
     this.nullableNumberSilly,
     this.nullableNumberSillySet,
   );
@@ -234,6 +239,8 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
   Set<TrivialNumber> numberSillySet;
 
   DateTime? dateTime;
+
+  TrivialString? trivialString;
 
   TrivialNumber? nullableNumberSilly;
   Set<TrivialNumber?> nullableNumberSillySet;