+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map__checked.dart
@@ -70,6 +70,7 @@ class _Factory implements k.KitchenSinkFactory<dynamic, dynamic> {
         TrivialNumber(0),
         {},
         DateTime.fromMillisecondsSinceEpoch(0),
+        TrivialString(''),
         TrivialNumber(0),
         {},
       );
@@ -195,14 +196,14 @@ class KitchenSink implements k.KitchenSink {
   }
 }
 
-@JsonSerializable(
-  checked: true,
-  anyMap: true,
-)
+@JsonSerializable(checked: true, anyMap: true, converters: [
+  // referencing a top-level field should work
+  durationConverter,
+  // referencing via a const constructor should work
+  BigIntStringConverter(),
+])
 // referencing a top-level field should work
-@durationConverter
-// referencing via a const constructor should work
-@BigIntStringConverter()
+@trivialStringConverter
 @TrivialNumberConverter.instance
 @EpochDateTimeConverter()
 class JsonConverterTestClass implements k.JsonConverterTestClass {
@@ -216,6 +217,7 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
     this.numberSilly,
     this.numberSillySet,
     this.dateTime,
+    this.trivialString,
     this.nullableNumberSilly,
     this.nullableNumberSillySet,
   );
@@ -238,6 +240,8 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
   Set<TrivialNumber> numberSillySet;
 
   DateTime? dateTime;
+
+  TrivialString? trivialString;
 
   TrivialNumber? nullableNumberSilly;
   Set<TrivialNumber?> nullableNumberSillySet;